+++ src/providers/twitch/eventsub/MessageBuilder.hpp
@@ -122,4 +122,16 @@ void makeModerateMessage(
     const lib::payload::channel_moderate::v2::Event &event,
     const lib::payload::channel_moderate::v2::AutomodTerms &action);
 
+/// <MODERATOR> initiated a raid to <CHANNEL>.
+void makeModerateMessage(
+    EventSubMessageBuilder &builder,
+    const lib::payload::channel_moderate::v2::Event &event,
+    const lib::payload::channel_moderate::v2::Raid &action);
+
+/// <MODERATOR> canceled the raid to <CHANNEL>.
+void makeModerateMessage(
+    EventSubMessageBuilder &builder,
+    const lib::payload::channel_moderate::v2::Event &event,
+    const lib::payload::channel_moderate::v2::Unraid &action);
+
 }  // namespace chatterino::eventsub